+++ core/src/com/orangelabs/rcs/core/ims/service/im/filetransfer/msrp/TerminatingMsrpFileSharingSession.java
@@ -270,7 +270,7 @@ public class TerminatingMsrpFileSharingSession extends ImsFileSharingSession imp
 
 			// Build SDP part
 	    	String ipAddress = getDialogPath().getSipStack().getLocalIpAddress();
-	    	int maxSize = ImsFileSharingSession.getMaxFileSharingSize();
+	    	long maxSize = ImsFileSharingSession.getMaxFileSharingSize();
 	    	String sdp = SdpUtils.buildFileSDP(ipAddress, localMsrpPort,
                     msrpMgr.getLocalSocketProtocol(), getContent().getEncoding(), fileTransferId,
                     fileSelector, null, localSetup, msrpMgr.getLocalMsrpPath(),
@@ -467,6 +467,7 @@ public class TerminatingMsrpFileSharingSession extends ImsFileSharingSession imp
      * @param currentSize Current transfered size in bytes
      * @param totalSize Total size in bytes
      * @param data received data chunk
+     * @return always true TODO
      */
     public boolean msrpTransferProgress(long currentSize, long totalSize, byte[] data) {
 		if (isSessionInterrupted() || isInterrupted()) {
